requirements.txt
streamlit
numpy
matplotlib
fpdf2
pandas
altair
scikit-learn
bcrypt
python-dotenv